+++ common/lib-sample-common/src/main/java/ru/surfstudio/android/sample/common/ui/base/loadstate/presentations/controllers/ErrorLoadStateController.kt
@@ -5,11 +5,9 @@ import ru.surfstudio.android.easyadapter.controller.NoDataItemController
 import ru.surfstudio.android.easyadapter.holder.BaseViewHolder
 import ru.surfstudio.android.sample.common.R
 
-
 /**
- * Контроллер для пустых состний в списках
- * */
-
+ * Controller for empty state in list
+ */
 class ErrorLoadStateController : NoDataItemController<ErrorLoadStateController.Holder>() {
 
     override fun createViewHolder(parent: ViewGroup) = Holder(parent)