+++ common/lib-sample-dagger/src/main/java/ru/surfstudio/android/sample/dagger/ui/base/dagger/screen/DefaultFragmentScreenModule.kt
@@ -42,16 +42,18 @@ class DefaultFragmentScreenModule(private val persistentScope: FragmentViewPersi
     @Provides
     @PerScreen
     internal fun provideEventDelegateManager(): ScreenEventDelegateManager {
-        return persistentScope.getScreenEventDelegateManager()
+        return persistentScope.screenEventDelegateManager
     }
 
     @Provides
     @PerScreen
-    internal fun providePermissionManager(eventDelegateManager: ScreenEventDelegateManager,
-                                          activityProvider: Provider<AppCompatActivity>,
-                                          activityNavigator: ActivityNavigator,
-                                          @Named(NO_BACKUP_SHARED_PREF) sharedPreferences: SharedPreferences,
-                                          fragmentProvider: FragmentProvider): PermissionManager {
+    internal fun providePermissionManager(
+            eventDelegateManager: ScreenEventDelegateManager,
+            activityProvider: Provider<AppCompatActivity>,
+            activityNavigator: ActivityNavigator,
+            @Named(NO_BACKUP_SHARED_PREF) sharedPreferences: SharedPreferences,
+            fragmentProvider: FragmentProvider
+    ): PermissionManager {
         return PermissionManagerForFragment(
                 eventDelegateManager,
                 activityProvider,
@@ -63,8 +65,10 @@ class DefaultFragmentScreenModule(private val persistentScope: FragmentViewPersi
 
     @Provides
     @PerScreen
-    internal fun provideMessageController(activityProvider: Provider<AppCompatActivity>,
-                                          fragmentProvider: FragmentProvider): MessageController {
+    internal fun provideMessageController(
+            activityProvider: Provider<AppCompatActivity>,
+            fragmentProvider: FragmentProvider
+    ): MessageController {
         return DefaultMessageController(activityProvider, fragmentProvider)
     }
 
@@ -76,16 +80,20 @@ class DefaultFragmentScreenModule(private val persistentScope: FragmentViewPersi
 
     @Provides
     @PerScreen
-    internal fun provideActivityNavigator(activityProvider: Provider<AppCompatActivity>,
-                                          fragmentProvider: FragmentProvider,
-                                          eventDelegateManager: ScreenEventDelegateManager): ActivityNavigator {
+    internal fun provideActivityNavigator(
+            activityProvider: Provider<AppCompatActivity>,
+            fragmentProvider: FragmentProvider,
+            eventDelegateManager: ScreenEventDelegateManager
+    ): ActivityNavigator {
         return ActivityNavigatorForFragment(activityProvider, fragmentProvider, eventDelegateManager)
     }
 
     @Provides
     @PerScreen
-    internal fun provideDialogNavigator(activityProvider: Provider<AppCompatActivity>,
-                                        fragmentProvider: FragmentProvider): DialogNavigator {
+    internal fun provideDialogNavigator(
+            activityProvider: Provider<AppCompatActivity>,
+            fragmentProvider: FragmentProvider
+    ): DialogNavigator {
         return DialogNavigatorForFragment(activityProvider, fragmentProvider, persistentScope)
     }
 }