+++ navigation/sample-standard/src/main/java/ru/surfstudio/android/navigation/sample_standard/screen/guide/GuideBindModel.kt
@@ -1,0 +1,12 @@
+package ru.surfstudio.android.navigation.sample_standard.screen.guide
+
+import ru.surfstudio.android.core.mvp.binding.rx.relation.mvp.Action
+import ru.surfstudio.android.dagger.scope.PerScreen
+import javax.inject.Inject
+
+@PerScreen
+class GuideBindModel @Inject constructor() {
+
+    val bottomNavClicked = Action<Unit>()
+    val sharedElementClicked = Action<Unit>()
+}