+++ navigation/sample/src/main/java/ru/surfstudio/android/navigation/sample/app/screen/splash/SplashFragment.kt
@@ -1,0 +1,26 @@
+package ru.surfstudio.android.navigation.sample.app.screen.splash
+
+import android.os.Bundle
+import android.os.Handler
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import ru.surfstudio.android.navigation.command.fragment.ReplaceHard
+import ru.surfstudio.android.navigation.sample.R
+import ru.surfstudio.android.navigation.sample.app.App
+import ru.surfstudio.android.navigation.sample.app.screen.auth.AuthRoute
+
+class SplashFragment : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_splash, container, false)
+    }
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        Handler().postDelayed({
+            App.navigator.execute(ReplaceHard(AuthRoute()))
+        }, 1500L)
+    }
+}