+++ push/sample/src/main/java/ru/surfstudio/android/push/sample/ui/screen/main/MainPresenter.kt
@@ -1,0 +1,22 @@
+package ru.surfstudio.android.push.sample.ui.screen.main
+
+import ru.surfstudio.android.core.mvp.presenter.BasePresenter
+import ru.surfstudio.android.core.mvp.presenter.BasePresenterDependency
+import ru.surfstudio.android.dagger.scope.PerScreen
+import javax.inject.Inject
+
+/**
+ * Презентер главного экрана
+ */
+@PerScreen
+internal class MainPresenter @Inject constructor(
+        basePresenterDependency: BasePresenterDependency
+) : BasePresenter<MainActivityView>(basePresenterDependency) {
+
+    private val sm: MainScreenModel = MainScreenModel()
+
+    override fun onLoad(viewRecreated: Boolean) {
+        super.onLoad(viewRecreated)
+        view.render(sm)
+    }
+}