+++ android-studio-settings/RELEASE_NOTES.md
@@ -1,7 +1,10 @@
-[TOC]
 # Android Studio Settings Release Notes
+
+- [0.5.0-alpha](#050-alpha)
+
 ## 0.5.0-alpha
 ##### Android-studio-settings
 * Init tracking changes in RELEASE_NOTES
 * Additional hunspell settings info.
 * ANDDEP-1021 Added FreeMarker template for MVI screen creation
+* ANDDEP-1085 settings.gradle have to update by hand for FeatureModule and InteractorModule. Temp hotfix because of android studio 4.0 bug