+++ common/RELEASE_NOTES.md
@@ -1,8 +1,13 @@
-[TOC]
 # Common Release Notes
+
+- [0.5.0-alpha](#050-alpha)
+- [Sample-common](#sample-common)
+
 ## 0.5.0-alpha
 ##### Sample-common
 * ANDDEP-785 added method scrollToBottom() in RecyclerViewUtils
 * ANDDEP-991 added method checkIfSnackbarIsVisible(String)
 * ANDDEP-978 Fix common dependency
 * ANDDEP-1038 Add PaginationFooterItemController for samples
+* ANDDEP-1037 Dependencies on deprecated module Network removed
+* ANDDEP-853 Translate docs