+++ core-mvp-binding/RELEASE_NOTES.md
@@ -1,5 +1,8 @@
-[TOC]
 # Core-mvp-binding Release Notes
+
+- [0.5.0-alpha](#050-alpha)
+- [0.3.0](#030)
+
 ## 0.5.0-alpha
 ##### Core-mvp-binding
 * ANDDEP-687 Added "androidx.constraintlayout:constraintlayout" dependency with "implementation" type
@@ -15,6 +18,7 @@ Added state checker methods to Request.
 `asRequest()` extensions moved to separate file `RequestRxExtension.kt`
 * ANDDEP-968 `Loading.kt`: added class `SimpleLoading`;
 * ANDDEP-968 `RequestUi.kt`: added fields `isLoading`, `hasData`, `hasError`;
+* ANDDEP-1048 Fixing wrong docs links and docs structure
 ## 0.3.0
 ##### Core-mvp-binding
 * Renamed `onViewDetached ()` -> `onViewDetach ()`