+++ mvi/RELEASE_NOTES.md
@@ -1,5 +1,7 @@
-[TOC]
 # Core-mvi Release Notes
+
+- [0.5.0-alpha](#050-alpha)
+
 ## 0.5.0-alpha
 ##### Core-mvi
 * ANDDEP-1008 License added
@@ -27,9 +29,14 @@ even if stream contains events from UI, which should be reacted directly.
 Replaced RxMiddleware.asRequestEvent `Event` parameter with `EventFactory`.
 * ANDDEP-968 `RequestEvent.kt`: added fields `isLoading`, `hasData`, `hasError`;
 * ANDDEP-1046 Fixed output events observing mechanism in CompositionTransformer: 
-now event composition mechanism supports middlewares that can only produces output events.  
+now event composition mechanism supports middlewares that can only produces output events.
+* ANDDEP-1048 Fixing wrong docs links and docs structure
 ##### Mvi-mapper
 * ANDDEP-968 **NEW** `RequestMapper.kt`: class for managing requests;
 * ANDDEP-968 **NEW** `RequestMapperLambdas.kt`: file that holds typealiases for `RequestMapper`;
+* ANDDEP-1048 Fixing wrong docs links and docs structure
 ##### Mvi-impls
 * ANDDEP-969 Added base classes: `BaseReactor` and `BaseReducer`;
+* **NO BACKWARD COMPATIBILITY** ANDDEP-1049 Package for `BaseActivityResultDelegate`,`SupportOnActivityResultRoute` and `CrossFeatureSupportOnActivityResultRoute` is changed 
+from `ru.surfstudio.android.core.ui.event.result` to:  `ru.surfstudio.android.core.ui.navigation.event.result`
+* ANDDEP-1048 Fixing wrong docs links and docs structure